LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/metadata/generated/window-location.https.sub-expected.txt


PASS sec-fetch-site - Same origin - location
PASS sec-fetch-site - Same origin - location.href
FAIL sec-fetch-site - Same origin - location.assign assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
FAIL sec-fetch-site - Same origin - location.replace assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
PASS sec-fetch-site - Cross-site - location
PASS sec-fetch-site - Cross-site - location.href
FAIL sec-fetch-site - Cross-site - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Cross-site - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-site - Same site - location
PASS sec-fetch-site - Same site - location.href
FAIL sec-fetch-site - Same site - location.assign assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same site - location.replace assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
PASS sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - location
PASS sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - location.href
FAIL sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect - location
PASS sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect - location.href
FAIL sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect - location.assign assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect - location.replace assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
PASS sec-fetch-site - Cross-Site -> Same Origin - location
PASS sec-fetch-site - Cross-Site -> Same Origin - location.href
FAIL sec-fetch-site - Cross-Site -> Same Origin - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Cross-Site -> Same Origin - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-site - Cross-Site -> Same-Site - location
PASS sec-fetch-site - Cross-Site -> Same-Site - location.href
FAIL sec-fetch-site - Cross-Site -> Same-Site - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Cross-Site -> Same-Site - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-site - Cross-Site -> Cross-Site - location
PASS sec-fetch-site - Cross-Site -> Cross-Site - location.href
FAIL sec-fetch-site - Cross-Site -> Cross-Site - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Cross-Site -> Cross-Site - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-site - Same-Origin -> Same Origin - location
PASS sec-fetch-site - Same-Origin -> Same Origin - location.href
FAIL sec-fetch-site - Same-Origin -> Same Origin - location.assign assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Same Origin - location.replace assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
PASS sec-fetch-site - Same-Origin -> Same-Site - location
PASS sec-fetch-site - Same-Origin -> Same-Site - location.href
FAIL sec-fetch-site - Same-Origin -> Same-Site - location.assign assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Same-Site - location.replace assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
PASS sec-fetch-site - Same-Origin -> Cross-Site - location
PASS sec-fetch-site - Same-Origin -> Cross-Site - location.href
FAIL sec-fetch-site - Same-Origin -> Cross-Site - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Cross-Site - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-site - Same-Site -> Same Origin - location
PASS sec-fetch-site - Same-Site -> Same Origin - location.href
FAIL sec-fetch-site - Same-Site -> Same Origin - location.assign assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Site -> Same Origin - location.replace assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
PASS sec-fetch-site - Same-Site -> Same-Site - location
PASS sec-fetch-site - Same-Site -> Same-Site - location.href
FAIL sec-fetch-site - Same-Site -> Same-Site - location.assign assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Site -> Same-Site - location.replace assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
PASS sec-fetch-site - Same-Site -> Cross-Site - location
PASS sec-fetch-site - Same-Site -> Cross-Site - location.href
FAIL sec-fetch-site - Same-Site -> Cross-Site - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same-Site -> Cross-Site - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - HTTPS downgrade-upgrade - location assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - HTTPS downgrade-upgrade - location.href assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - HTTPS downgrade-upgrade - location.assign assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - HTTPS downgrade-upgrade - location.replace assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-mode - location
PASS sec-fetch-mode - location.href
PASS sec-fetch-mode - location.assign
PASS sec-fetch-mode - location.replace
PASS sec-fetch-dest - location
PASS sec-fetch-dest - location.href
PASS sec-fetch-dest - location.assign
PASS sec-fetch-dest - location.replace
PASS sec-fetch-user - location
PASS sec-fetch-user - location.href
PASS sec-fetch-user - location.assign
PASS sec-fetch-user - location.replace
FAIL sec-fetch-user - location with user activation assert_own_property: expected property "sec-fetch-user" missing
FAIL sec-fetch-user - location.href with user activation assert_own_property: expected property "sec-fetch-user" missing
FAIL sec-fetch-user - location.assign with user activation assert_own_property: expected property "sec-fetch-user" missing
FAIL sec-fetch-user - location.replace with user activation assert_own_property: expected property "sec-fetch-user" missing
